requirements.txt
PyQt6>=6.5.0
python-dotenv
pypdf
markdown
requests
pillow
